Release 09/2025

+++ app/lib/util/temporary_storage_handler.dart
@@ -21,18 +21,29 @@ class TemporaryStorageHandler {
     return "user-id_${_userId}_study-id_${_studyId}_$timestamp";
   }
 
-  static Future<Directory> _getMultimodalTempDirectory() async {
-    final tempAppData = await getTemporaryDirectory();
-    final multimodalTempDirectory =
-        Directory("${tempAppData.path}/multimodal-temp");
-    await multimodalTempDirectory.create(recursive: true);
-    return multimodalTempDirectory;
+  static Future<Directory?> _getMultimodalTempDirectory() async {
+    try {
+      final tempAppData = await getTemporaryDirectory();
+      final multimodalTempDirectory = Directory(
+        "${tempAppData.path}/multimodal-temp",
+      );
+
+      if (!await multimodalTempDirectory.exists()) {
+        await multimodalTempDirectory.create(recursive: true);
+      }
+
+      return multimodalTempDirectory;
+    } catch (e) {
+      StudyULogger.error(e);
+      return null;
+    }
   }
 
   static Future<Directory> _getMultimodalUploadDirectory() async {
     final appData = await getApplicationDocumentsDirectory();
-    final multimodalUploadDirectory =
-        Directory("${appData.path}/multimodal-upload");
+    final multimodalUploadDirectory = Directory(
+      "${appData.path}/multimodal-upload",
+    );
     await multimodalUploadDirectory.create(recursive: true);
     return multimodalUploadDirectory;
   }
@@ -43,14 +54,7 @@ class TemporaryStorageHandler {
   ) async {
     final stagingFile = File(stagingFilePath);
     final uploadDirectory = await _getMultimodalUploadDirectory();
-    final uploadFile = File(
-      path.join(
-        uploadDirectory.path,
-        [
-          blobId,
-        ].join(),
-      ),
-    );
+    final uploadFile = File(path.join(uploadDirectory.path, [blobId].join()));
     await stagingFile.rename(uploadFile.path);
   }
 
@@ -64,8 +68,13 @@ class TemporaryStorageHandler {
     return futureBlobFiles;
   }
 
-  Future<FutureBlobFile> getStagingAudio() async {
+  Future<FutureBlobFile?> getStagingAudio() async {
     final temporaryMultimodalDirectory = await _getMultimodalTempDirectory();
+
+    if (temporaryMultimodalDirectory == null) {
+      return null;
+    }
+
     final fileName = _buildFileName();
     final localFilePath = path.join(
       temporaryMultimodalDirectory.path,
@@ -75,13 +84,21 @@ class TemporaryStorageHandler {
         TemporaryStorageHandler._audioFileType,
       ].join(),
     );
-    final futureBlobId =
-        [fileName, TemporaryStorageHandler._audioFileType].join();
+
+    final futureBlobId = [
+      fileName,
+      TemporaryStorageHandler._audioFileType,
+    ].join();
     return FutureBlobFile(localFilePath, futureBlobId);
   }
 
-  Future<FutureBlobFile> getStagingImage() async {
+  Future<FutureBlobFile?> getStagingImage() async {
     final temporaryMultimodalDirectory = await _getMultimodalTempDirectory();
+
+    if (temporaryMultimodalDirectory == null) {
+      return null;
+    }
+
     final fileName = _buildFileName();
     final localFilePath = path.join(
       temporaryMultimodalDirectory.path,
@@ -91,21 +108,31 @@ class TemporaryStorageHandler {
         TemporaryStorageHandler._imageFileType,
       ].join(),
     );
-    final futureBlobId =
-        [fileName, TemporaryStorageHandler._imageFileType].join();
+
+    final futureBlobId = [
+      fileName,
+      TemporaryStorageHandler._imageFileType,
+    ].join();
     return FutureBlobFile(localFilePath, futureBlobId);
   }
 
   static Future<void> deleteAllStagingFiles() async {
     final temporaryMultimodalDirectory = await _getMultimodalTempDirectory();
-    for (final file in await temporaryMultimodalDirectory
-        .list()
-        .where(
-          (f) => path
-              .basename(f.path)
-              .startsWith(TemporaryStorageHandler._stagingBaseNamePrefix),
-        )
-        .toList()) {
+
+    if (temporaryMultimodalDirectory == null) {
+      StudyULogger.error("Temporary multimodal directory is null");
+      return;
+    }
+
+    for (final file
+        in await temporaryMultimodalDirectory
+            .list()
+            .where(
+              (f) => path
+                  .basename(f.path)
+                  .startsWith(TemporaryStorageHandler._stagingBaseNamePrefix),
+            )
+            .toList()) {
       await file.delete();
     }
   }
